+++ packages/terra-docs/src/terra-dev-site/tool/example/CautionNotice.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Notice } from '@cerner/terra-docs';
+
+export default () => (
+  <Notice variant="caution" ariaLevel="3">
+    This component was created for this express purpose and caution should be used when exceeding those parameters.
+  </Notice>
+);