+++ packages/terra-docs/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Unreleased
 
+* Introducing the button, placeholder and notice components.
 * Initial stable release